Update classes.py
StaffMembers are now stored as instances, not just as dictionaries with their attributes.

We now assign capabilities to each staff member, as well as required capabilities to each component.

The program assigns staff members to tasks according to their capabilities and the remaining balance of work days. The program ensures that the amount of work performed by each staff member does not exceed an overall workday_cap.